Rearrange routes to accomodate project EDITS
This change required some rewiring of the route structure so that templates were not nesting incorrectly.
