+++ app/components/project-form.js
@@ -1,0 +1,155 @@
+import Component from '@ember/component';
+import { action, computed } from '@ember-decorators/object';
+import mapboxgl from 'mapbox-gl';
+import { service } from '@ember-decorators/service';
+import { argument } from '@ember-decorators/argument';
+import { tagName } from '@ember-decorators/component';
+
+const developmentSiteLayer = {
+  type: 'line',
+  paint: {
+    'line-color': 'green',
+    'line-width': 4,
+  },
+};
+
+const projectAreaLayer = {
+  type: 'line',
+  paint: {
+    'line-color': 'orange',
+    'line-width': 4,
+  },
+};
+
+const rezoningAreaLayer = {
+  type: 'line',
+  paint: {
+    'line-color': 'purple',
+    'line-width': 4,
+  },
+};
+
+const selectedLotsLayer = {
+  type: 'fill',
+  paint: {
+    'fill-color': 'rgba(217, 216, 1, 1)',
+    'fill-outline-color': 'rgba(255, 255, 255, 1)',
+  },
+};
+
+@tagName('')
+export default class ProjectFormComponent extends Component {
+  constructor(...args) {
+    super(...args);
+    this.set('selectedLots', {
+      type: 'FeatureCollection',
+      features: [],
+    });
+  }
+
+  @service
+  notificationMessages;
+
+  @service
+  router;
+
+  @argument
+  model
+
+  developmentSiteLayer = developmentSiteLayer
+
+  projectAreaLayer = projectAreaLayer
+
+  rezoningAreaLayer = rezoningAreaLayer
+
+  selectedLotsLayer = selectedLotsLayer
+
+  lotSelectionMode = false
+
+  projectGeometryMode = null
+
+  isSelectingLots = false
+
+  @computed('selectedLots.features.[]')
+  get selectedLotsSource() {
+    const selectedLots = this.get('selectedLots');
+    return {
+      type: 'geojson',
+      data: selectedLots,
+    };
+  }
+
+  @action
+  handleSearchSelect(result) {
+    const map = this.get('mapInstance');
+
+    // handle address search results
+    if (result.type === 'lot') {
+      const center = result.geometry.coordinates;
+      this.set('searchedAddressSource', {
+        type: 'geojson',
+        data: {
+          type: 'Feature',
+          properties: {},
+          geometry: result.geometry,
+        },
+      });
+
+      if (map) {
+        map.flyTo({
+          center,
+          zoom: 15,
+        });
+      }
+    }
+  }
+
+  @action
+  handleSearchClear() {
+    this.set('searchedAddressSource', null);
+  }
+
+  @action
+  handleMapLoad(map) {
+    this.set('mapInstance', map);
+    window.map = map;
+
+    // setup controls
+    const navigationControl = new mapboxgl.NavigationControl();
+
+    map.addControl(navigationControl, 'top-left');
+  }
+
+  @action
+  handleLayerClick(feature) {
+    const { id: layerId } = feature.layer;
+
+    // if lot was clicked when in lot selection mode, handle the clicl
+    if (layerId === 'pluto-fill' && this.get('lotSelectionMode')) {
+      const { type, geometry, properties } = feature;
+      const selectedLots = this.get('selectedLots');
+
+      // if the lot is not in the selection, push it, if it is, remove it
+      const inSelection = selectedLots.features.find(lot => lot.properties.bbl === properties.bbl);
+
+      if (inSelection === undefined) {
+        this.get('selectedLots.features').pushObject({
+          type,
+          geometry,
+          properties,
+        });
+      } else {
+        this.set('selectedLots.features', selectedLots.features.filter(lot => lot.properties.bbl !== properties.bbl));
+      }
+    }
+  }
+
+  @action
+  async save(model) {
+    const project = await model.save();
+
+    this.get('notificationMessages').success('Project saved!');
+
+    this.get('router').transitionTo('projects.show', project);
+  }
+}
