+++ app/routes/projects/edit/index.js
@@ -1,7 +1,18 @@
 import Route from '@ember/routing/route';
+import { action } from '@ember-decorators/object';
+import { service } from '@ember-decorators/service';
 
-export default class ProjectsEditIndexRoute extends Route {
+export default class ProjectsEditRoute extends Route {
+  @service
+  notificationMessages;
+
   model() {
     return this.modelFor('projects.edit');
+  }
+
+  @action
+  error({ message }) {
+    this.get('notificationMessages').error(message);
+    this.transitionTo('application');
   }
 }
