+++ app/templates/components/project-form.hbs
@@ -110,5 +110,3 @@
     {{/labs-map}}
   {{/project-map-form}}
 </div>
-
-{{outlet}}
