+++ app/templates/projects/edit/index.hbs
@@ -1,84 +1,3 @@
-<div class="cell">
-  <div class="grid-x grid-padding-x grid-padding-y">
-    <div class="cell medium-4">
-
-      <h1>Project Name</h1>
-      <ul>
-        <li><strong>Applicant:</strong> Jane Jacobs</li>
-        <li><strong>Tax Lots:</strong> Queens Block 02293, Lot 1; Block 02294 32, Lots 30, 31, 32, 33</li>
-      </ul>
-      <p><img src="http://via.placeholder.com/600x400?text=MAP" /></p>
-      <button class="button expanded gray">Edit Project</button>
-
-    </div>
-    <div class="cell auto">
-
-      {{!-- Existing maps --}}
-      <h3 class="">Existing Maps</h3>
-      <p>Choose one of your existing project maps to view, print, or edit.</p>
-      <div class="grid-x grid-padding-x medium-up-3">
-        {{#each model.applicantMaps as |map|}}
-          <div class="cell">
-            {{#link-to 'projects.edit.map.new' map tagName='div' class='card link text-center' role='button'}}
-              <img src="/img/sample-map-1.gif" alt="Sample Map" />
-              <div class="card-section">
-                {{map.id}}
-              </div>
-            {{/link-to}}
-          </div>
-        {{/each}}
-      </div>
-
-      <hr>
-
-      {{!-- New Maps --}}
-      <h3 class="">Add New Map</h3>
-      <p>Choose the type of map you would like to add to your project.</p>
-      <div class="grid-x grid-padding-x medium-up-2 large-up-4">
-        <div class="cell">
-          {{#link-to 'projects.edit.map.new' model.id tagName='div' class='card link text-center'}}
-            <div class="card-divider bg-green-muted green-light align-center">
-              {{fa-icon 'plus' size='3x' class=''}}
-            </div>
-            <div class="card-section">
-              Area
-            </div>
-          {{/link-to}}
-        </div>
-        <div class="cell">
-          <div class="card link text-center">
-              <div class="card-divider bg-green-muted green-light align-center">
-                {{fa-icon 'plus' size='3x' class=''}}
-              </div>
-            <div class="card-section">
-              Tax
-            </div>
-          </div>
-        </div>
-        <div class="cell">
-          <div class="card link text-center">
-              <div class="card-divider bg-green-muted green-light align-center">
-                {{fa-icon 'plus' size='3x' class=''}}
-              </div>
-            <div class="card-section">
-              Zoning Change
-            </div>
-          </div>
-        </div>
-        <div class="cell">
-          <div class="card link text-center">
-              <div class="card-divider bg-green-muted green-light align-center">
-                {{fa-icon 'plus' size='3x' class=''}}
-              </div>
-            <div class="card-section">
-              Zoning Section
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-
-</div>
+{{project-form model=model}}
 
 {{outlet}}
