+++ app/templates/projects/new.hbs
@@ -1,114 +1,3 @@
-<div class="cell medium-5 large-3 sidebar">
-  <label class="text-small">Project Title
-    {{input type="text" class='project-name-field' value=model.projectName}}
-  </label>
-
-  <label class="text-small">Applicant Name
-    {{input type="text" class='applicant-name-field' value=model.applicantName}}
-  </label>
-
-  <label class="text-small">ZAP Project ID
-    {{input type="text" class='zap-project-id-field' value=model.zapProjectId}}
-  </label>
-
-  <label class="text-small">Description
-    {{textarea class='zap-project-description-field' value=model.description rows="3"}}
-  </label>
-
-  <label class="text-small">Project Geometries</label>
-
-  {{#if mapInstance}}
-    {{draw-control
-      model=model
-      mapInstance=mapInstance
-      mode="developmentSite"
-      modeDisplayName="Development Site"
-      projectGeometryMode=projectGeometryMode
-      selectedLots=selectedLots
-      lotSelectionMode=lotSelectionMode
-    }}
-
-    {{draw-control
-      model=model
-      mapInstance=mapInstance
-      mode="projectArea"
-      modeDisplayName="Project Area"
-      projectGeometryMode=projectGeometryMode
-      selectedLots=selectedLots
-      lotSelectionMode=lotSelectionMode
-    }}
-
-    {{draw-control
-      model=model
-      mapInstance=mapInstance
-      mode="rezoningArea"
-      modeDisplayName="Rezoning Area"
-      projectGeometryMode=projectGeometryMode
-      selectedLots=selectedLots
-      lotSelectionMode=lotSelectionMode
-    }}
-  {{/if}}
-
-  <button type="button" class="button expanded project-save-button" disabled={{not model.isValid}} onclick={{action 'save' model}}>Create Project</button>
-
-</div>
-<div class="cell medium-7 large-9 map-container">
-  {{#project-map-form tagName='' as |projectMapForm|}}
-    {{#labs-map
-      id='main-map'
-      initOptions=(hash
-        style=projectMapForm.model.meta.mapboxStyle
-        zoom=12
-        center=(array -73.983307 40.704977)
-      )
-      mapLoaded=(action handleMapLoad) as |map|}}
-
-        {{!-- labs-layers (includes tax-lots layer) --}}
-        {{map.labs-layers
-          layerGroups=projectMapForm.model.layerGroups
-          onLayerClick=(action 'handleLayerClick')
-        }}
-
-        {{!-- Ad-hoc source and layer for developmentSite --}}
-        {{#if model.developmentSite}}
-          {{#map.source options=(mapbox-geojson-source model.developmentSite) as |source|}}
-            {{source.layer
-              layer=developmentSiteLayer
-              before='boundary_country'
-            }}
-          {{/map.source}}
-        {{/if}}
-
-        {{!-- Ad-hoc source and layer for projectArea --}}
-        {{#if model.projectArea}}
-          {{#map.source options=(mapbox-geojson-source model.projectArea) as |source|}}
-            {{source.layer
-              layer=projectAreaLayer
-              before='boundary_country'
-            }}
-          {{/map.source}}
-        {{/if}}
-
-        {{!-- Ad-hoc source and layer for projectArea --}}
-        {{#if model.rezoningArea}}
-          {{#map.source options=(mapbox-geojson-source model.rezoningArea) as |source|}}
-            {{source.layer
-              layer=rezoningAreaLayer
-              before='boundary_country'
-            }}
-          {{/map.source}}
-        {{/if}}
-
-
-        {{!-- Ad-hoc source and layer for showing selected lots in lot selection mode --}}
-        {{#map.source options=selectedLotsSource as |source|}}
-          {{source.layer
-            layer=selectedLotsLayer
-            before='boundary_country'
-          }}
-        {{/map.source}}
-    {{/labs-map}}
-  {{/project-map-form}}
-</div>
+{{project-form model=model}}
 
 {{outlet}}
