+++ mirage/config.js
@@ -17,6 +17,7 @@ export default function() {
 
   this.get('/projects');
   this.get('/projects/:id');
+  this.patch('/projects/:id');
   this.post('/projects');
   this.post('/applicant-maps');
   /*
